requirements.txt
# Core LLM and AI libraries
langchain-nvidia-ai-endpoints>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain>=0.2.0

# Neo4j and graph database
neo4j>=5.14.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# NLP and text processing
spacy>=3.7.0
python-dotenv>=1.0.0

# HTTP and API clients
requests>=2.31.0

# Data validation
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6

# Voice I/O
elevenlabs>=1.0.0

# Audio processing
sounddevice>=0.4.6
soundfile>=0.12.1
pyaudio>=0.2.13

# Jupyter notebook
jupyter>=1.0.0
ipywidgets>=8.0.0
matplotlib>=3.7.0

# Streamlit web interface
streamlit>=1.28.0
streamlit-chat>=0.1.1
streamlit-extras>=0.3.0
